Add deploy target to Makefile
Using `make deploy` will cause the site source to be built and then invoke
the `nikola github_deploy` command which updates the master branch with the
contents of the output/ directory and pushes to the remote master branch.

The ghp-import tool does much of the above. The `nikola github_deploy`
wrapper is used because it generates a useful message for the master
commit.

N.B. Although `make deploy` causes the master branch to be pushed, the src
branch must be committed and pushed separately.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>