Use Hugo static site generator
Plus updated content, README.md, Makfile, etc.

Also, instead of relying on a peer stgit repository to be present when
building the static site, the sync'ed artifacts from the stgit repo (man
pages, changelog) are now checked-in to this website repo. This enables
the site to be built and deployed with more independance, which is aids
in publishing using a GitHub Action.

N.B. Hugo templates are used to fix-up links in the man pages instead of
using asciidoc macros in an asciidoc config file. Hugo's model for invoking
asciidoc is not amenable to either the customized config or command line
options used in the stgit repo. See layouts/man/single.html.